// frontend/front-web/src/pages/Membro/detalhe-membro/types.tsx
export type MembroDetail ={
    id: number;
    nome: string;
    sobrenome: string;
    email: string;
    idade: number;
    dataNascimento: Date;
    telefone:  string;
    url: string;
    cpf: string;
    estadoCivil: number;
    pequenoGrupo: pequenoGrupo;
}
export type pequenoGrupo={
    id:number;
    apelido: string;
}